Move _AssertLogsContext in a private unittest._log submodule.
